Remove config option `editor.backUpBeforeSaving`
After shipping #11828 Atom always creates a backup file in the
`~/.atom/recovery` directory, trying to restore it automatically via the
main process in case there is a hard crash on the renderer process. This
approach should be resilient enough to allow us to delete this config
option altogether.

We can still keep it for a while in text-buffer, in case we need to use
it again at some point, but it's probably fine to remove it from there
too at some point in the future.