Always copy item when splitting from the context menu
This updates the context menu so that splitting behaves consistently,
regardless of whether the context menu is used from a text editor or
not.

The current behavior comes from #7953 but, as @maxbrunsfeld pointed out
when we talked about this, the purpose of that commit was to simplify
the command names, so the UX change probably wasn't intentional.