Make main-process/main.js as minimal as possible
In order to test changes to main.js, we need to rebuild the application.
For this reason, it makes sense to do as little work as possible here.
In this commit, we change main to just determine the resourcePath, then
use that path to require a start function which can be iterated on in
dev mode.

Signed-off-by: Max Brunsfeld <maxbrunsfeld@github.com>