Call debounced cursor blink resume callback any time the cursor moves
Previously, we were only calling the debounced resume callback when the
cursor was already blinking, but this meant we didn't renew the debounce
window. This in turn meant we attempted to start blinking the cursor
again every 100ms regardless of cursor interactions, which was consuming
an extra frame and causing cursor movement to not feel smooth.