Don't snapshot minimatch and fix package transpilation registry on win32
This module uses Node's `path` for determinining which path separator to
use on the current platform. On browsers (and every other environment
that does not support `require`, such as v8 snapshots) it falls back to
always using a forward slash.

As a result, `PackageTranspilationRegistry` (and potentially other
bundled packages that depend on `minimatch`) couldn't match glob
expressions against any given path on Windows, thus causing the custom
transpiler code to not work properly.