Always add opened items to the workspace
This commit changes the behavior of `open()` so that the item will be
added to the workspace—even if the `activateItem` option is false.

This provides a way to open (and add) an item without activating it,
but removes the ability to open without adding. For that, we've made
the `createItemForURI()` method public 😊