Require `TextEditorElement` eagerly
With snapshots, all the forbidden modules are required lazily. In the
case of `TextEditorElement` this can be problematic because users might
create editors via `document.createElement('atom-text-editor')`.

With this commit we will eagerly require
`src/text-editor-element.coffee`, thus forcing `TextEditorElement` to
register the HTML custom element and fixing the issue.