Fix regexps for excluding C++ sources from the binary
Previously, a spurious '*' in these regexps would cause them to match anything in
the containing modules' src folder. This caused problems for modules like
oniguruma that have JavaScript files in the src folder.

Also, we were shipping superstring's C++ code because we forgot to add it to
the blacklist. I think it's safe to say we don't want to ship any C++ code
at all.