:bug: Add missing onDidAddTextEditor method to Dock
Dock's observeTextEditors() method calls this.onDidAddTextEditor(), but
Dock didn't have an onDidAddTextEditor method. :scream_cat:

Dock's observeTextEditors() method also calls this.getTextEditors(), and
Dock's getTextEditors() method calls
this.paneContainer.getTextEditors(), but there is no getTextEditors()
method on this.paneContainer. :see_no_evil:

This commit adds a test for observeTextEditors, and it adds the missing
onDidAddTextEditor method, and it fixes the getTextEditors method to
make the new test pass.